[Ecommerce][BE] Refactor login and payment flow

+++ src/main/java/com/hcmut/ecommerce/domain/payment/dto/request/CreateEscrowRequest.java
@@ -8,7 +8,7 @@ import lombok.Data;
 @Builder
 @AllArgsConstructor
 public class CreateEscrowRequest {
-    private Float amount;
+    private Integer amount;
     private String buyerId;
     private String sellerId;
     private String orderId;
